refactor: Move HomeTemplate into /templates/home folder
